Core/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
websocket-client>=1.8.0
